# swagger-spec.json
{"openapi":"3.0.0","paths":{"/hello":{"get":{"operationId":"HelloController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["hello"]}},"/public/activity/list":{"get":{"operationId":"ActivityController_findAllPaginated","parameters":[{"name":"order","required":false,"in":"query","schema":{"default":"ASC","enum":["ASC","DESC"],"type":"string"}},{"name":"page","required":false,"in":"query","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"take","required":false,"in":"query","schema":{"minimum":1,"maximum":50,"default":10,"type":"number"}}],"responses":{"200":{"description":"Successfully received model list","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/PageDto"},{"properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/ResponseActivityDto"}}}}]}}}}},"tags":["activity"]}},"/public/activity/all":{"get":{"operationId":"ActivityController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ResponseActivityDto"}}}}}},"tags":["activity"]}},"/public/activity/{id}":{"get":{"operationId":"ActivityController_findOneById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseActivityDto"}}}}},"tags":["activity"]},"put":{"operationId":"ActivityController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateActivityDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseActivityDto"}}}}},"tags":["activity"]},"delete":{"operationId":"ActivityController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseActivityDto"}}}}},"tags":["activity"]}},"/public/activity":{"post":{"operationId":"ActivityController_save","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateActivityDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseActivityDto"}}}}},"tags":["activity"]}},"/public/address/{id}":{"get":{"operationId":"AddressController_findOneById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseAddressDto"}}}}},"tags":["address"]},"put":{"operationId":"AddressController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAddressDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseAddressDto"}}}}},"tags":["address"]},"delete":{"operationId":"AddressController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseAddressDto"}}}}},"tags":["address"]}},"/public/address":{"post":{"operationId":"AddressController_save","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAddressDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseAddressDto"}}}}},"tags":["address"]}},"/public/cabinet/list":{"get":{"operationId":"CabinetController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ResponseCabinetDto"}}}}}},"tags":["cabinet"]}},"/public/cabinet/{id}":{"get":{"operationId":"CabinetController_findOneById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseCabinetDto"}}}}},"tags":["cabinet"]},"put":{"operationId":"CabinetController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCabinetDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CabinetEntity"}}}}},"tags":["cabinet"]},"delete":{"operationId":"CabinetController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CabinetEntity"}}}}},"tags":["cabinet"]}},"/public/cabinet":{"post":{"operationId":"CabinetController_save","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCabinetDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseCabinetDto"}}}}},"tags":["cabinet"]}},"/public/country/list":{"get":{"operationId":"CountryController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ResponseCountryDto"}}}}}},"tags":["country"]}},"/public/country/{id}":{"get":{"operationId":"CountryController_findOneById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseCountryDto"}}}}},"tags":["country"]}},"/public/currency/list":{"get":{"operationId":"CurrencyController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ResponseCurrencyDto"}}}}}},"tags":["currency"]}},"/public/currency/{id}":{"get":{"operationId":"CurrencyController_findOneById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseCurrencyDto"}}}}},"tags":["currency"]}},"/public/tax/list":{"get":{"operationId":"TaxController_findAll","parameters":[{"name":"order","required":false,"in":"query","schema":{"default":"ASC","enum":["ASC","DESC"],"type":"string"}},{"name":"page","required":false,"in":"query","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"take","required":false,"in":"query","schema":{"minimum":1,"maximum":50,"default":10,"type":"number"}}],"responses":{"200":{"description":"Successfully received model list","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/PageDto"},{"properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/TaxEntity"}}}}]}}}}},"tags":["tax"]}},"/public/tax/{id}":{"get":{"operationId":"TaxController_findOneById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["tax"]},"put":{"operationId":"TaxController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTaxDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["tax"]},"delete":{"operationId":"TaxController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["tax"]}},"/public/tax":{"post":{"operationId":"TaxController_save","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTaxDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["tax"]}}},"info":{"title":"Invoice System API Server APIs Specification","description":"Section for describe whole APIs","version":"1.0","contact":{}},"tags":[],"servers":[{"url":"http://0.0.0.0:8081/","description":"Local environment"},{"url":"https://staging.yourapi.com/","description":"Staging"},{"url":"https://production.yourapi.com/","description":"Production"}],"components":{"securitySchemes":{"accessToken":{"scheme":"bearer","bearerFormat":"JWT","type":"http"},"refreshToken":{"scheme":"bearer","bearerFormat":"JWT","type":"http"},"apiKey":{"type":"apiKey","in":"header","name":"x-api-key"},"permissionToken":{"type":"apiKey","in":"header","name":"x-permission-token","description":"grant permission for /admin prefix endpoints"}},"schemas":{"PageMetaDto":{"type":"object","properties":{"page":{"type":"number"},"take":{"type":"number"},"itemCount":{"type":"number"},"pageCount":{"type":"number"},"hasPreviousPage":{"type":"boolean"},"hasNextPage":{"type":"boolean"}},"required":["page","take","itemCount","pageCount","hasPreviousPage","hasNextPage"]},"PageDto":{"type":"object","properties":{"data":{"type":"array","items":{"type":"array"}},"meta":{"$ref":"#/components/schemas/PageMetaDto"}},"required":["data","meta"]},"ResponseActivityDto":{"type":"object","properties":{"deletedAt":{"format":"date-time","type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"id":{"type":"number","example":1},"label":{"type":"string","example":"Service"}},"required":["deletedAt","createdAt","updatedAt","id","label"]},"CreateActivityDto":{"type":"object","properties":{"label":{"type":"string","minLength":3,"example":"monetize frictionless paradigms"}},"required":["label"]},"UpdateActivityDto":{"type":"object","properties":{"label":{"type":"string","minLength":3,"example":"monetize frictionless paradigms"}},"required":["label"]},"ResponseCountryDto":{"type":"object","properties":{"deletedAt":{"format":"date-time","type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"id":{"type":"number","example":1},"name":{"type":"string","example":"Spain"}},"required":["deletedAt","createdAt","updatedAt","id","name"]},"ResponseAddressDto":{"type":"object","properties":{"deletedAt":{"format":"date-time","type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"id":{"type":"number","example":1},"address":{"type":"string","example":"5889 Chandler Ridges"},"address2":{"type":"string","example":"Apt. 509"},"region":{"type":"string","example":"New Zachary"},"zipcode":{"type":"string","example":"98241-1922"},"country":{"$ref":"#/components/schemas/ResponseCountryDto"}},"required":["deletedAt","createdAt","updatedAt","id","address","address2","region","zipcode","country"]},"CreateAddressDto":{"type":"object","properties":{"address":{"type":"string","example":"82641 Field Lane"},"address2":{"type":"string","example":"Apt. 855"},"region":{"type":"string","example":"Fort Carmelocester"},"zipcode":{"type":"string","example":"23960"},"countryId":{"type":"number","example":1}},"required":["address","address2","region","zipcode","countryId"]},"UpdateAddressDto":{"type":"object","properties":{"address":{"type":"string","example":"82641 Field Lane"},"address2":{"type":"string","example":"Apt. 855"},"region":{"type":"string","example":"Fort Carmelocester"},"zipcode":{"type":"string","example":"23960"},"countryId":{"type":"number","example":1}},"required":["address","address2","region","zipcode","countryId"]},"ResponseCurrencyDto":{"type":"object","properties":{"deletedAt":{"format":"date-time","type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"id":{"type":"number","example":1},"label":{"type":"string","example":"Barbados Dollar"},"code":{"type":"number","example":"KMF"},"symbol":{"type":"number","example":"$"}},"required":["deletedAt","createdAt","updatedAt","id"]},"ResponseCabinetDto":{"type":"object","properties":{"deletedAt":{"format":"date-time","type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"enterpriseName":{"type":"string","example":"Lowe Group"},"email":{"type":"string","example":"Arthur.Miller-Zieme23@hotmail.com"},"phone":{"type":"string","example":"471.510.9340 x93642"},"taxIdNumber":{"type":"string","example":"564500075"},"address":{"$ref":"#/components/schemas/ResponseAddressDto"},"activity":{"$ref":"#/components/schemas/ResponseActivityDto"},"currency":{"$ref":"#/components/schemas/ResponseCurrencyDto"}},"required":["deletedAt","createdAt","updatedAt","address"]},"CreateCabinetDto":{"type":"object","properties":{"enterpriseName":{"type":"string","example":"Satterfield LLC"},"email":{"type":"string","example":"Anastacio31@hotmail.com"},"phone":{"type":"string","example":"640.631.5477"},"taxIdNumber":{"type":"string","minLength":9,"example":"375569166"},"address":{"$ref":"#/components/schemas/CreateAddressDto"},"activityId":{"type":"number","example":1},"currencyId":{"type":"number","example":1}},"required":["address"]},"UpdateCabinetDto":{"type":"object","properties":{"enterpriseName":{"type":"string","example":"Satterfield LLC"},"email":{"type":"string","example":"Anastacio31@hotmail.com"},"phone":{"type":"string","example":"640.631.5477"},"taxIdNumber":{"type":"string","minLength":9,"example":"375569166"},"address":{"$ref":"#/components/schemas/CreateAddressDto"},"activityId":{"type":"number","example":1},"currencyId":{"type":"number","example":1}},"required":["address"]},"ActivityEntity":{"type":"object","properties":{"id":{"type":"number"},"label":{"type":"string"}},"required":["id","label"]},"CurrencyEntity":{"type":"object","properties":{"id":{"type":"number"},"label":{"type":"string"},"code":{"type":"string"},"symbol":{"type":"string"}},"required":["id","label","code","symbol"]},"CountryEntity":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"}},"required":["id","name"]},"AddressEntity":{"type":"object","properties":{"id":{"type":"number"},"address":{"type":"string"},"address2":{"type":"string"},"region":{"type":"string"},"zipcode":{"type":"string"},"country":{"$ref":"#/components/schemas/CountryEntity"},"countryId":{"type":"number"}},"required":["id","address","address2","region","zipcode","country","countryId"]},"CabinetEntity":{"type":"object","properties":{"id":{"type":"number"},"enterpriseName":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"taxIdNumber":{"type":"string"},"activity":{"$ref":"#/components/schemas/ActivityEntity"},"activityId":{"type":"number"},"currency":{"$ref":"#/components/schemas/CurrencyEntity"},"currencyId":{"type":"number"},"address":{"$ref":"#/components/schemas/AddressEntity"},"addressId":{"type":"number"}},"required":["id","enterpriseName","email","phone","taxIdNumber","activity","activityId","currency","currencyId","address","addressId"]},"CreateTaxDto":{"type":"object","properties":{"label":{"type":"string","minLength":3,"example":"FODEC"},"rate":{"type":"number","maximum":1,"minimum":1,"example":"0.05"},"isSpecial":{"type":"boolean","example":"true"}},"required":["label","rate","isSpecial"]},"UpdateTaxDto":{"type":"object","properties":{"label":{"type":"string","minLength":3,"example":"FODEC"},"rate":{"type":"number","maximum":1,"minimum":1,"example":"0.05"},"isSpecial":{"type":"boolean","example":"true"}},"required":["label","rate","isSpecial"]}}}}